adicionando views

+++ EcoWave/Repositories/ItemRecicladoRepository.cs
@@ -1,11 +1,8 @@
-﻿using EcoWave.Data;
-using EcoWave.Models;
-using EcoWave.Repositories;
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
 using EcoWave.Data;
 using EcoWave.Models;
 using Microsoft.EntityFrameworkCore;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 
 namespace EcoWave.Repositories
 {
@@ -20,12 +17,12 @@ namespace EcoWave.Repositories
 
         public async Task<IEnumerable<ItemReciclado>> GetAll()
         {
-            return await _context.ItensReciclados.ToListAsync();
+            return await _context.ItensReciclados.Include(i => i.Usuario).ToListAsync();
         }
 
         public async Task<ItemReciclado> GetById(int id)
         {
-            return await _context.ItensReciclados.FindAsync(id);
+            return await _context.ItensReciclados.Include(i => i.Usuario).FirstOrDefaultAsync(i => i.ItemId == id);
         }
 
         public async Task Add(ItemReciclado entity)
@@ -42,9 +39,12 @@ namespace EcoWave.Repositories
 
         public async Task Delete(int id)
         {
-            var entity = await _context.ItensReciclados.FindAsync(id);
-            _context.ItensReciclados.Remove(entity);
-            await _context.SaveChangesAsync();
+            var entity = await GetById(id);
+            if (entity != null)
+            {
+                _context.ItensReciclados.Remove(entity);
+                await _context.SaveChangesAsync();
+            }
         }
     }
 }
